+++ resources/js/Views/Guide/Show.vue
@@ -35,3 +35,4 @@ export default {
     props: ['content'],
 }
 </script>
+