+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Bill actions adhere to specialized contracts
 - Updated documentation
 - Code cleanup
+- Rate limiter works only in production
+
+### Fixed
+
+- `ParsesBillInformation` is not instantiable
 
 ## [v1.0.1 (2020-06-13)](https://github.com/emberfuse/scorch/compare/v1.0.0...v1.0.1)